New mostly-integer-only sound mixer
The mixer itself is integer-only, but the millibel stuff is still
float-based (I don't know if it's worth adding a 10001-value-long
lookup table to replace it).

Anyway, according to a quick test, this new mixer is significantly
faster than the old floating-point one - rarely going above 100
clock() ticks per callback, with the old one almost always running
above that.

For now, only the SDL2 backend supports it - I'll add the others in
upcoming commits.

This will no-doubt cause problems with the enhanced (lite) branch,
where clownaudio always outputs float samples.

+++ src/Backends/Audio/SoftwareMixer.h
@@ -13,4 +13,4 @@ void Mixer_RewindSound(Mixer_Sound *sound);
 void Mixer_SetSoundFrequency(Mixer_Sound *sound, unsigned int frequency);
 void Mixer_SetSoundVolume(Mixer_Sound *sound, long volume);
 void Mixer_SetSoundPan(Mixer_Sound *sound, long pan);
-void Mixer_MixSounds(float *stream, unsigned int frames_total);
+void Mixer_MixSounds(short *stream, unsigned int frames_total);
